.emacs.el: dired-goto-last-file: don't redisplay
I originally added a forced `redisplay` here so
that the window would always end up scrolled to
show the empty line under the last Dired entry.

This was a higher-level UX issue that I was
coupling to this function for convenience - it
was practice writing "good enough" code, and
much lighter on time and energy than thinking
of the best way to factor it otherwise.

At the time it also seemed that this was
Always the Objectively Correct Behavior.

I had a use-case in going to the end of my
note list, and initially found myself
reflexively wondering if there was another
line below the one I am looking at.

But now a year later, I didn't even "know"
I was seeing an extra blank line at the
bottom. I forgot and didn't notice. If you
had asked me, I might've confidently said
my code doesn't bother to add that nicety.
Either it became purely subconscious, or I
built up enough trust/assumption that my
my code is showing me the last files with
nothing below.

This is a great reminder that some UI
feedback might have value that's relative
to the user's familiarity/confidence/trust
in the tool.

So I'm going to try dropping this behavior
and see if I start doubting/checking if
there's another note at the bottom of the
note list. And if I don't, then this is a
trust-relative UI feedback (at least for
me), or something in that vein.