.emacs.el: evil-dired-toggle-hidden: better positioning
When toggling from showing to hiding, if
point is on a hidden file, it won't have a
file-name to return to, and that leaves
the general-purpose position-preserving
logic with only line numbers to work with.

But in the toggle-hidden code we know we are
about to hide files (and we are already in
the business of knowing some definition of
what makes a file "hidden"): so we can just
check if we are currently on a hidden file,
and move to the nearest/next non-hidden file,
so that point ends up on a sensible line after
the hiding/revert.

This commit implements a basic version of
that. (I chose to only look downwards for the
next non-hidden file, rather than upwards or
both, because it makes for simpler prediction
as a user: I don't have to know/notice if
there is a nearer file in the other direction
to predict where I will go (and there is a
kinetic skeumorphism and thus cognition reuse
to going down rather than up - if the file
I'm on disappears, I will drop as if due to
gravity, or like in a platformer).)