.emacs.el: task: drop dired-subdir-alist hack
Now that the top header line is once again
being hidden and not deleted, worth trying
to delete this again - we don't need this
position hack with narrowing, we do need
it with deletion. Off the top of my head
we don't need it with invisible+intangible,
but I can't rule out eventually tripping
over some code that does the wrong thing
thing in the presence of those properties.