Move task list filtering out of Dired
* Add an `emacs-task-ls` script, which can be
  used as the `insert-directory-program` in
  lisp instead of `ls` when getting the task
  list. About as "minimally viable" as it gets.

* Rewrite the task-listing stuff in `.emacs.el`
  to use this new script.

First impression is that it works without any
issues.

The big practical benefit is that it moves all
the filtering inside of Dired's core logic -
and thus automatically benefits from all of the
position-related code. It's hard enough getting
the position-preserving and scroll-adjusting
logic right when reverting. This way we don't
have to implement additional scroll-fixing
logic to compensate for filtering after the
Dired revert.