.emacs.el: without-local-variable
It turns out, that some code misbehaves in
tricky ways when buffer-local and lexical
scopes interact.

For example, when calling `dired`, we can
`let`-bind `process-environment` to pass
custom environment variables to
`insert-directory-program` (usually
`ls`). But if we do that in an Eshell
buffer, it doesn't work - turns out,
Eshell makes `process-environment` a
local variable, and `kill-local-variable`
is sufficient to make it work again.

So this commit provides a macro that will
kill one or more local variables (if those
symbols are in fact local), executes body,
and then restores them (makes them local
if they were local, and restores their
local value if they had one).