.emacs.el: drop delete-{file,directory} hack
This is now more correctly moved into Dired's
functions. The idea is that code should know
when it does a file operation if it might be
part of a user's actions, or if it's always a
low-level mechanical thing that never is - in
the latter cade, it's reasonable to optimize
by calling a low-level delete that doesn't
integrate with Dired; otherwise code should
be calling deletion functions which have an
opportunity to update UI if desired.