.emacs.el: give up on fake-fringe for now
It's a lot nicer than no fringe, but SO MUCH
code in Emacs ecosystem struggles with it. It's
ridiculous. So much time and energy spent, and
no end in sight to various subtle bugs.

(Most recently, I've noticed that my REPL
wrapping bugs out around the screen edge,
which I'm pretty sure is a new issue. I used
that stuff daily for months and of course
implemented it to my high standards. Now
suddenly it starts bugging out in a new way?
fake-fringe is the biggest candidate, and is
the only relevant change I remember making
or installing. There's already at least two
issues with Eat being off-by-one due to
line-prefix/wrap-prefix. The last straw that
led me to this commit was noticing that some
weird jumped-to-the-wrong-place behavior in
`less` (f.e. `git log` in popped Eshell+Eat)
was fixed by disabling fake fringe.)