.emacs.el: histdir-repl: disable buffer-undo-list
It doesn't work in Eat buffers/terminals
anyway, but leaving it enabled causes any
undo keybind to do minor junky edits to the
buffer. And in vi-style UX it's very easy
to accidentally tap undo ("u" is wrapped
on all but the top with frequently used
keys) - especially on a phone touch screen.