.emacs.el: music--line-move-after-refresh
A special line-move/evil-line-move,
which takes effect at the next redraw.

This is needed for vi-like "o" and "p"
to work properly, most noticeably at
the bottom of the buffer, because the
lines of added files aren't actually
inserted into the buffer until the next
refresh, so the desired line and column
might not be available until then.