.emacs.el: music: fix "o"/"p" move edge-cases
This fixes edge-cases at the bottom of the
buffer and when pasting enough lines to
span/displace the currently playing song's
lines. For example, `p` or `o` at the very
end of the buffer (no song on current line)
will now place point correctly on the last
song (and this is the base case, when first
adding files to a new/empty music buffer).