.emacs.el: smoother-dired: only scroll/pulse if found
For example, if invoked from a file such as
`.emacs.el`, and hidden files are not being
shown, this would recenter on and pulse some
other entry.

I'm not sure if this is actually preferable
UX, but it seems worth trying - because the
pulse is reliably spotted peripherally, the
absense of a noticed pulse might be learned
as a reliable cue that there's not even any
need to look around for point. (The downside
is that the eyesight damage that accumulates
with age can eventually add up to false
negatives with that.)