.emacs.el: add-dired-entry: fix error on bad path
Behave like `dired-add-{file,entry}`, and
just silently do nothing when given a path
that's not found in the buffer.

In particular, this includes paths that start
with "~/", so it's debatable if this is the
right move. These `add-dired-*` wrappers are
explicitly heavier-weight convenience wrappers.
The big DX benefit of expanding the path is
that it's hard to remember/predict when some
API might return paths unexpanded. So I think a
good case exists for automatically expanding
file names in this function. But not
automatically expanding is consistent with most
of Dired's other functions, so code code which
has to call other Dired functions already has
to expand the path anyway... and so I'm holding
off on it for now.