.emacs.el: smoother-dired: ensure source entry
Add entry of the source file (if smoother-dired
was called from a file) if it's not already
present.

One of the big benefits of smoother-dired is
that it allows quickly doing any Dired action
from inside the file you currently have open.

(For example, "Space d x x" is an
extremely fast way to get to an Eshell
in the current directory with the
current file's path pre-filled, and
instead of being a special-purpose
command, it's the composition of "Space
d" for `smoother-dired` and then `x`
for my `evil-dired-eshell` operator.)

Ensuring that the source file is actually
present if possible makes such UX flows a
lot more reliable. (In particular, this
frees thinking and muscle memory from being
predicated on state like "does the Dired
buffer already exist?" "is it up-to-date?"
"is this file hidden?" "is the Dired buffer
showing hidden files?")