.emacs.el: shred: stop interpreting nil as 1
`shred` handling nil `register` argument as 1
was left over from v1 of the shred concept,
where pasting immediately shredded. (vi/Evil
paste commands have a good reason to allow a
nil register argument as the default.)

Now that `shred` is a separate action from
pasting and `shred-candidate` is initialized
to 1, `shred-candidate` is always (supposed
to be) a register, and there's no more reason
to specially tolerate nil as an alias for 1.