.emacs.el: music: set "^" vi mark on current song
This was one idea that resulted from realizing
that it would be nice to have a motion to the
currently playing thing.

vi's "^" mark fits fairly well semantically,
because my music player's analog to vi "insert"
mode is "seek" mode: seeking is always done in
the currently playing file; entering seek mode
places cursor on the currently playing file.