.emacs.el: music-play: unpause when switching
The main motivation is that playing a song from
a paused-and-cleared state is the same UX (same
keypresses) as playing a song from a freshly
created or cleared-without-pausing state.

Another benefit is smoother UX when the player
is paused and the desire is to start playing
another song. (Switching without unpausing is
a much rarer use-case, and can now be achieved
by hitting Enter twice.)

(This also means that pausing through this
command is more deterministic: hitting it
twice on a song (other than current) will
end with player being paused, regardless
of current pause state.)