.emacs.el: read-buffer-multiple: return buffers instead of buffer names
Generally the righter API, and fixes a class of
bugs for all callers: if you loop on the result
list and take actions which might cause other
buffers to be renamed (such as killing buffers
of files with the same name), buffer objects
remain valid but buffer names are cached info
that might be invalidated.

For example, when using `smoother-kill-buffer`,
if `README.md<dir1>` and `README.md<dir2>` are
chosen, then killing one will cause the other
to be renamed to `README.md` before the second
kill is attempted.