.emacs.el: histdir-repl: PyREPL line-wrap support
This uses the line-wrap-suffix feature added in
the last commit to support PyREPL-based REPLs.

(For now, I'm just hard-coding this to activate
for all Python REPLs. That's technically wrong,
but the new logic can't misbehave in another
Python REPL implementation unless that REPL also
actually outputs line feeds in the middle of
single-line inputs.)