.emacs.el: stop using my `p` script
Android app Emacs has full native clipboard
integration, so I can just do normal paste
on all of my systems now.

This also eliminates the significant Termux API
delay (~1s of waiting for it to spin up just to
grab the clipboard, during which Termux must
remain as the foreground app with input focus).
Finally!

(In theory I could still need `p` when running
Emacs in terminals, so there's an argument for
keeping the `p`-using path in the code, but I
am optimistically betting it will never come
up, because terminal Emacs is only a fallback
when the GUI isn't working, which I haven't
needed to use in the two years since switching
to Emacs as my daily-driver editor/everything.)