.emacs.el: smoother-fill-paragraph: fix last line
When measuring jaggedness, don't ignore the last
line if it juts out further than the rest of the
paragraph (after ignoring overhang characters).

A paragraph ending with a line that comes far
short of its visual right edge looks normal. A
paragraph ending with a line that sticks out
past its visual right edge is odd.

Note that

* when the paragraph has only two lines,
  this precludes reflows which make the
  second line longer than the first, but

* in all other cases, the last line can*
  *longer than the others, it just has to
  *"earn" its excess length by the same
  *rules as the other lines.