.emacs.el: lambda-let: fix binding
This fixes commit 08f66f474ef3ba722e88d1ce144bd0a11548f1c2.

That previous commit broke `lambda-let` in
contexts where lexical binding is on, such as
the minibuffer during `eval-expression` and
buffers which have `lexical-binding` set to
`t`. (Which is most of modern Emacs lisp.)

Basically, it turns out that when `eval` is
invoked inside lexical bindings and is told
to use lexical bindings, the form inside
`eval` does not inherit any of the lexical
bindings from the caller of `eval`.

Luckily, if lexical binding is already
on, we don't need to call `eval`.

The tricky part is that for correctness
in all cases:

1. we must wait to get the value of
   `lexical-binding` when the macro's
   expansion is actually being evaluated
   - that way, it works even if the
   macro is expanded in one buffer, and
   then the expanded form is evaluated
   in another buffer with a different
   `lexical-binding` value; AND YET

2. we must get the value of `lexical-binding`
   through a macro, because that variable is
   special - if you evaluate `lexical-binding`
   directly inside the minibuffer of a command
   like `eval-expression`, you will get its
   value for the buffer the command is about,
   not the value for the current minibuffer -
   but for whatever reason, if you evaluate it
   while expanding a macro, you always get the
   value which is currently in effect for the
   current buffer.