.emacs.el: prevent select of inactive echo area
On Termux, touches of the inactive minibuffer
have started to select that "echo area" window.
I suspect this was added in Emacs 30.

(Almost no keybinds are bound there, because
it's not supposed to be selectable. So the only
way to get out is to tap elsewhere, and reaching
up with the thumb to do that is less ergonomic
and slower than any keyboard-driven option.)