.emacs.el: no dialog boxes by default
It's nice that the option/capability to pop
native GUI questions is available. In
particular, it would be good when Emacs is
backgrounded. But when I'm in Emacs and
already have the keyboard directly under my
fingers, a GUI dialog that blocks/ignores
the keyboard and requires me to reach and
paw at a specific apot in the middle of the
screen is... not good.