.emacs.el: Android: exclusively use Termux PATH
Android app Emacs, unlike Termux, includes
various system paths in PATH. Since those
directories do are locked down to the point
of not letting programs get a list of what's
in there, Tab-completion in Eshell breaks if
those paths are in there. In any case, Termux
seems to provide everything we need.