.emacs.el: Android key translations in fancy input
Make the key translations from the last commit
work even when the Android input method is
doing its own processing instead of passing
through raw input.

(This code was actually done May 4th or so,
but it didn't make sense to commit it until I
was happy enough with my keybindings to commit
those.)

Partially broken by the fact that pushing the
decoded events back on the unread events list
causes the events to be translated a second
time, resulting in recursive translations
(f.e. `¹` -> `1` -> `Escape`). (When this
code was written, I was only remapping the
long presses.)