.emacs.el: work around Android frame width issue
During `after-make-frame-functions`, Android
Emacs

* reports the correct width for the first
  frame, but then

* reports an incorrect, much larger width
  (both in pixels and in columns).

This causes my font size adjustment to
only be applied on the first frame.

Once I even saw new buffers in the initial
frame start getting created with font size 11.5
instead of 11.0. If I remember right, this
happened after using another app. I suspect
this is due to Android-specific interactions
(such as Android killing an Activity that's
backgrounded - maybe Emacs then creates a new
frame for the same activity).

This commit works around that by recording
the width of the initial frame, and using
that for the check whenever a new frame is
being created.