.emacs.el: exif: orientation misparse workaround
`exif.el` mis-parses the orientation tag in some
images, for example reading `0x80000` instead of
`8` in some photos I have (taken by Android in
early 2022, with an up-to-date Pixel 5, and then
rotated in the default picture viewer on an
up-to-date Windows 11 in early/mid 2025).

This workaround maps the mis-parsed codes to the
same definitions as the correct codes.