git-pop: add `--continue` option
Nice when the pop causes a merge conflict.

Ideally this would also drop the entry that
was kept by the failed pop, to truly finish
everything that the interactive pop started,
but that's tricky in the general case (to
continue `git-pop <n>`, we need to remember
<n> somewhere and parse it back, to do `git
stash drop <n>`; but then also the stash
could've gained new entries since then or
been otherwise changed, so we also need
`git-pop --continue <n>` as an escape hatch).