.emacs.el: pixel-scroll-align-edge
When pixel-precise scrolling is in use,
the top and bottom lines might be only
partially in view.

For some reason, when point is in one of those
lines (and thus only partially visible), Emacs
does not automatically scroll those lines into
view.

This commit adds that behavior as a global
post-command hook.