.emacs.el: datetime-read preview: ensure resize
On Android Emacs (and possibly GUI Emacs
frames in other graphical environments with
narrow-enough screens), the datetime preview
ends up not growing the minibuffer when it
spills onto a second line. This results in
the actual input being invisible and only
the last characters of the parsed datetime
or error being visible.

I didn't see this while using Termux (Emacs'
minibuffer resize probably works differently
in terminal frames, since that whole logic
has to be different).