.emacs.el: dired-binary-search
Allow bisecting a Dired buffer (assuming it is
sorted) to find the first entry which does not
sort before the wanted entry (which will be the
wanted entry if the wanted entry exists, or the
line at which to insert the wanted entry if it
needs to be added).

If the Dired buffer's entries are not sorted
lexically, it is the caller's responsibility
to specify the correct `predicate`/comparator
function. (A mature solution would provide a
helpet function which returns the comparator
that matches the sort implied by the `ls`
switches used, and uses that as the default
`predicate` value, but for now I don't need
this.)