.emacs.d/init.el: fix font sizing on Pixel 5
The current workaround that adjusts the font
size on Pixel 5 only works if I don't switch
away from the Emacs app before it renders.

When the frame initializes while it's
out-of-view, it has some size that it
will never actualy render with - on
Pixel 5 it's 82 characters wide, on
Pixel 6 it's 83.

This commit switches to a workaround that's
more hard-coded to Pixel 5's quirks, but will
cover this case reliably.

(A more ideal solution would probably
use `window-size-change-functions`.)