.emacs.d/init.el: " %": random-line
Add a function and keybinding to jump to a
random line. Sometimes comes up even in text
files, but more importantly, in a paradigm
where everything can be represented as text,
it allows picking at random from any list of
items - for example, pick a song at random
in Dired or in my music player.