.emacs.d/init.el: fix for my Vertico resizing
When the last character of input is at the
right edge of the minibuffer, the cursor gets
wrapped to the next line. This expands the
minibuffer to have one more vertical line for
input, but that line is empty, so it's not
counted by `count-screen-lines`, and as a
result the bottom line of the candidate list
will be pushed out of view.