.emacs.d/init.el: note: improved paste UX
When creating a note from a pasticle, leave
point at the start of the note body, and make
sure there's only one blank line between the
front matter and the pasted content and a
trailing newline after (this was not happening
when pasting Evil's linewise yanks; similarly,
Evil's blockwise yanks no longer paste an
extra blank column).

Creating a note from a linewise copy is
very common for me, and usually I want to
do some editing at the start of the note
(add a title, a prompt for an AI, or an
opening code block fence), and moving to
the end is a lot nicer than moving to the
start of the note ("G" vs having to move
between the frontmatter and body).

Note: using `evil-paste-before` only achieves
the correct behavior because I have that
function monkeypatched to behave the way I
that consider correct/proper/better.