Emacs: interpreted-function-lisp helper function
Translates the latest representation of
non-compiled closures into the standard
and stable lisp form: `(lambda ...)`.

When the closure has lexical bindings,
the returned form is wrapped to include
the bindings: `(let (...) (lambda ...))`.

The fourth field (index 3) is ignored
because it is always nil for interpreted
functions (it is only used for compiled
functions, to declare stack size).