Emacs: Eat: fix line-wrap logic
Patch `eat--t-write` to correctly compute when
it needs to line-wrap terminal output. Biggest
benefit is that this fixes a corner case where
a multi-column character (CJK, emoji, etc) that
falls on the end of the line causes Eat to lock
up in an infinite loop.

The actual logic fix is courtesy of Yikai Zhao:
https://codeberg.org/akib/emacs-eat/pulls/228#issuecomment-7382149