Emacs: fix for my Vertico resizing (part 2)
The last fix relied on the cursor being at the
end of the line. (This was handled in a prior
alternative fix approach, but got lost in the
simplifying refactor.)

The last fix also only worked on the first line
wrap (when the minibuffer had fully filled only
only line, as opposed to fully filling two or
more).