Emacs: enhanced CRM: magic backspace; drop ";"
Alternative solution to the case of not
wanting to continue selecting within the
same narrowed candidate set:

Immediately after hitting ",", Backspace
will clear the current narrowing input (the
part of the input after the CRM separator,
which the "," reinserts). Any other action
will cause Backspace to work normally again.

This optimizes the most common case, which is
a habit of hitting "," whenever you want to
add the current candidate to your choices,
without the mental overhead of ahead-of-time
predication on whether this will consume the
current narrowing input (if there is only one
candidate left in the current narrowed set)
and whether you want that or not. (This habit
is the natural default because it does the
desired thing a lot of the time, costs less
mentally, and is quicker. Having an extra ";"
binding and actually choosing whether to hit
"," or ";" adds the overhead of a decision
point -- if both keys were equally accessible
then it might still be worth the learning
curve, since it's not exactly a decision one
the intent is mapped directly to muscle
memory, but on GBoard (and most mobile
keyboards), ";" is a long-press.)