Emacs: Android: fix for virtual directories
The directories

* `/`
* `/assets`, and
* `/content`

on Android's Emacs are fake mappings to
contents inside Emacs' APK. Since they
cannot be accessed through the actual
file system, trying to `ls` them fails
with an error. (A non-obvious error,
because Dired code was written to just
hide the error instead of grabbing it
and including it in the error message.)

This commit re-enables using lisp instead of
an external `ls` program in the default case
on Android, leaving external program use for
just my Task buffers.