Emacs: destructive-completion: change highlight
Use a different highlight face in minibuffers
that are running a completion command that is
destructive or otherwise worth calling out.

I'm only using+testing this with Vertico, but
I've written it in a way that should work
with icomplete, and in principle any other
completion UIs that use the `highlight` face
inside the minibuffer.