Emacs: count-vertico-input-lines improvement
Use the existing overlay (and the temporary
mutation of it that my code is already
doing) to force `count-screen-lines` to get
the correct amount. Having exactly one fake
character at the end of the line forces the
line count to correctly handle the edge
case when the end is at exactly the line
edge, since the extra character line-wraps
same as the cursor itself would.

(Technically this fix might be wrong if the
window is ever one character wide, but at
that point everything's busted and unusable,
and that's an improvement over `string-width`
being wrong if `vertico-count-format` ever
contains a tab character or whatever.)

(We still need the `count-final-newline`
option to `count-screen-lines`. I thought
this was because the overlay is using
`before-string`, but placing the placeholder
character into `after-string` didn't seem to
eliminate this need, it's probably subtler
than that -- may even be a bug in how
something in Emacs handles the combination
of trailing newline and trailing overlay.)

(I would also bet this performs better,
since it does less work in Lisp and
leans more onto one pass of Emacs'
native rendering and line-counting).