Emacs: Vertico: per-command counter faces
Allow remapping `vertico-counter` face for
specific commands. Whether this ought to be
in Vertico is more debatable than just the
base face itself, but it still clearly fits
the primary justification for including a
thing in a module rather than externally:
there are more than one independent use-cases
for this (or would be, given enough vision).