Emacs: reorder some lines
When I started using `setq-if-nil` everywhere
rather than avoiding it in all functions that
were useful outside of my config, I forgot to
move it up earlier in my config, before any of
the uses.

Otherwise, the exact place to put it is
fairly arbitrary, so I put it right after the
`apply-split` macro that it uses, which also
groups it at the start of my other assignment
macros, such as `seq-setq*` and `uncons`.