Emacs: enhanced CRM: support multichar separator
Properly handle `crm-separator` being
multiple characters.

This still does not handle the case where the
crm-separator is a regular expression, but it
does get the code most of the way there: now
it just needs to insert a string that would
be matched by `crm-separator` rather than
inserting the the separator as-is.