Emacs: enhanced CRM: use ",\n" as separator
Improve readability of candidates by having
each candidate on its own line(s).

Unlock UX efficiencies and versatility from
being able to navigate and manipulate chosen
candidates as lines of text. For example:

* "dd" to unselect a candidate;

* "yy" and "p" to repeat a selection.

Also improve interoperability, (consistency
of) readability outside of the minibuffer, and
gain UX efficiency, by once again using a
literal comma. For example:

* CRM inputs once again survive round trips
  through other programs/systems;

* all the normal vi motions work correctly
  again ("f"/"F"/"t"/"T"; word boundary
  detection matches what's displayed).

(Using a real comma is possible again because
a two-character separator, especially when one
of those characters is a newline, is nearly
perfectly unambigious and trivially escapable.
Any reasonably accessible character by itself
was a gaping pothole in the space of
likely-to-encounter candidates, that could not
be escaped without either look-behind support
in Emacs' regular expressions or an support in
CRM for splitting on a subgroup of the
separator. But once there's two characters, we
can just split them with a backslash if we
ever need to match that lair in a candidate.)