Emacs: enhanced CRM: add newline to separator
Improve readability of candidates by having
each candidate on its own line(s).

Unlock UX efficiencies and versatility from
being able to navigate and manipulate chosen
candidates as lines of text. For example:

* "dd" to unselect a candidate;

* "yy" and "p" to repeat a selection.

As an unexpected bonus, this solves the edge
case of candidates being pushed out of view
once accumulated choices grow the minibuffer
to its limit. (When choices are all on one
logical line, the line-wrapped line gets
precedence over the candidate overlay; when
the choices are separate lines, Emacs just
pushes the top lines out of view, letting
the candidate overlay stay fully in view.
Presumably the candidate overlay is counted
as part of the last line, so when there's
multiple lines, Emacs can use its model of
which logical lines are supposed to be in
view.)