Emacs: smoother-kill-buffer: make API "safer"
When invoked as a *command* by a human, it
makes sense to kill the current buffer.

But when invoked as an *API* by code, taking
`nil` as a default is an opening for a bug
elsewhere in the system to turn into silently
wrong behavior, in this case destructively
so. For example, when a completion framework
is implemented such that `require-match` is
effectively ignored by CRM, and `get-buffer`
returns `nil` is used as a "does not exist"
result.