Emacs: denoted-tag-edit
Replaces/upgrades `denoted-tag-add` on "gh".

For adding tags, this is still the best
UX path.

For editing tags and reordering tags,
this adds a best UX path.

For removing tags along with doing any of the
above at once, this can sometimes be the best
UX path.

For just removing tags, `denoted-tag-remove`
on "gr" is still the best UX path.