Emacs: dired-abbreviate-headers: fix rename case
Use an overlay with a display property instead
of a text replacement. Dired code relies on the
full expanded path to identify that the Dired
buffer needs to be renamed when the directory
is renamed.

Also make sure the overlay is (re-)applied
as needed when the buffer is renamed.