Emacs: smoother-fill-paragraph: soft-minimum-fill-column
Allow more tightening if it leads to a
bigger jaggedness win.

For example, in my previous commit, we have:

> Make renames of Dired buffers that are
> visiting renamed directories uniquify
> the same way that is done for renames
> of buffers that are visiting files.

Before this change, with a 40-column minimum,

> 1234567890123456789012345678901234567890

that would've filled to

> Make renames of Dired buffers that are
> visiting renamed directories uniquify
> the same way that is done for renames of
> buffers that are visiting files.

, because the narrowing was stopped just before
it could squeeze the "of". But in this case, it
is visually worth it, and numerically, a -1 to
the fill column wins -2 variance (and an even
bigger win to the other metrics of jaggedness
that I've entertained).