Emacs: smoother-delete-file: y-or-n for kill
The marginal safety benefit of confirm-p over
y-or-n-p doesn't apply as much to the second
prompt: the first is likely to have already
blocked any "auto-pilot" keypresses, and to
have gotten the user's attention.

Conversely, I do find myself wanting the
ergonomics of hitting "n" for the second
prompt, when wanting to delete the file but
not yet kill the buffer.