Emacs: music: optimize current entry face use
Instead of applying a different face on each
redraw (and awkwardly passing the face two
layers down the stack), have the line always
propertized with the "current entry" face,
and use face remaps to make Emacs work.

I'm undecided if this is the "proper way",
but it seems nicer to factor it in a way
that decouples playing/paused state from
modifying the buffer.