Emacs: music: "d": pull point up to last line
When deleting the last lines of a buffer,
point will now be moved up to the last line
that remains, instead of being left at the
bottom of the buffer (past the last newline).

The check for `>=` instead of `=` isn't just
best practice here, it turns out to actually
make a correctness difference -- without it,
I saw semi-reliably-reproducible cases where
the cursor failed to jump up. (Primarily, by
hitting "d" quickly, so that the 2nd "d" of
"dd" and the 1st "d" of the next "dd" were
close enough together, although I also saw
this happen without getting into "operator"
state, and logically it doesn't quite make
sense for keypress processing to cause this.
Instead, what I think is happening is that
the timer that redraws the buffer sometimes
fires *between* the delete IPC call to `mpv`
and the `(point-max)` call, so the latter
returns the max *after* deletion, making
this a great example of why and how it's
best practice to cover all values *beyond*
the expected bound, even if they "should be
impossible".)