Emacs: music: make playing line an overlay
It's more intuitive for vi-style motions
when this line isn't counted.

1. You are normally thinking at the abstract
   level of songs, not lines, so consciously
   adjusting the vi count to account for the
   playing line becomes inefficient UX (when I
   first wrote this UX, it was more intuitive
   for me to think and move in raw lines, but
   that inflected eventually).

2. Having the current line be a true line can
   be confusing. If you cut/delete/copy the
   current song on the playing line, your
   pasticle is that line: a "current / total"
   time and no visible song information. I
   just had a moment where this confused me
   while shredding clipboard entries -- I had
   to me to pause and retrace my steps because
   I thought I was seeing the result of a bug.

The downside is that a textual copy no longer
preserves playing information -- which song
was playing, position in the song, looping or
not -- and that information could've been used
to automatically restore that state upon paste
into the music buffer. But in practice, I have
never missed that, whereas I *have* wanted the
playing line to become an overlay.