Emacs: music: put face back into buffer
Having it in the overlay was convenient and
technically more efficient, but currently
causes `describe-face` to not offer it.

More generally, this is a good reminder that
other code which might want to inspect or
manipulate faces will have a harder time with
it when it's "out of the way" in an overlay.

I had been on the fence about putting it
there anyway -- the efficiency/simplicity
of just adjusting an overlay was exciting,
but it felt a little wrong semantically.

Also, if I want this buffer to behave correctly
with `display-line-numbers` (show the line
number after current song on the next song's
line rather than showing it too early, on the
play line), then I need to make the playing
line a `display` property that's replacing just
the newline, and then the file line's face
needs to either be a different overlay or part
of the buffer.