Emacs: music: remove unused music-open & " sm"
Haven't used this in months, maybe closing on
a year -- basically ever since I wrote the
vi-style controls. "Space m o" is the same
number of keystrokes as "Space s m", and if I
need any thought or precision, I can do other
actions in the music buffer before hitting
"o".

The new extra motivation for removing this
is that I'm adding undo/redo support, and
this function would complicate everything
since it alone bypasses my vi-style code
and paradigm (`music--open`; getting index
from under point in music buffer).

The one use-case that remains for the " sm"
keybinding is to enqueue/search music without
switching to the music buffer at all, but at
this point if I want that I will reimplement
it on top of the vi-style open functions.