Emacs: music: use relative seeking
This might fix the bug where the point keeps
jumping back sometimes. Near as I've been
able to tell, whenever this bug happens, the
underlying position is almost at the top of
the next second -- I haven't consciously
diagnosed it beyond that. One idea I had was
that `mpv` might be rejecting seeks that are
too small (especially because absolute seeks
are exact by default).