Emacs: music: "c"/"C": get names before deleting
Now that there's redraw batching, the
insertion logic needs to always happen
once the deletion happens, since they
expect to be paired (most prominently,
for batching redraws -- a delete which
thinks it's paired with an insertion
won't redraw).

This change is just one way to do it: prompt
for the names before doing the deletion, so if
the user aborts the prompt, they are cancelling
the whole operation and there's no deletion.

(Arguably, it's more vi-like to do the delete
and redraw immediately, and then just edit the
delete's undo entry from unpaired to paired
once the insertion happens. But as far as UX
semantics/intuition/expectations go, I am used
to minibuffer prompts happening before the
operation and being a "nevermind" opportunity.)

(The full vi analog would be to do the deletion,
drop the user into Insert mode, let the user
trigger completion of file names in-buffer with
Tab or whatever, and enqueue the tracks and
amalgamate history upon escape to Normal mode.)