Emacs: latest-eshell: fix cycling getting stuck
Ensure that `latest-eshell` called from an
Eshell always chooses the latest among any
other Eshells that exist.

This was already the intended behavior but the
fallback for new windows missed this case, and
a window's previous buffer list can end up with
the current Eshell buffer as the latest Eshell
entry when switching to a non-Eshell buffer and
back.