Emacs: music-play: handle end-of-buffer case
When hitting Enter in the empty line past the
end of the list, clear the currently selected
index (set it back to -1).

Mostly this is just polishes the UX to not
error, but it's also the most sensical thing
to do in this case, and a natural way to make
this state transition accessible to the user.