Emacs: text-property-values: refactor
Use my `setq-if-nil` here instead of `unless`
because the semantics is "default value".

I find that clearer, easier to mentally parse.

Also, by setting the default for `end` as
well as `start`, the loop arguably becomes
clearer, and we move the branch out of the
loop (but modern optimization passes know
how to do that).