Emacs: evil-dired: internal cleanup
I guess I want to transition off my habit
of using `dired-directory` in cases where
it would be wrong if it contains wildcards.

Originally the idea was that it's clearer if
we just assume (as in all my uses so far) that
it's a directory, but I find myself no longer
wanting to be coupled to that assumption,
while being more comfortable with expecting
that `default-directory` will never deviate
from the correct directory for the purposes
of functions like this.